Create ShortestPathThroughMaze.py
A solution to a very interesting problem:
Find the shortest path in a maze from start to finish.
Input: A rectangular matrix of 1 and 0.
  1: represents a wall
  0: represents a corridor
There is the following extra rule: You can remove 
exactly one wall -- i.e. you can turn a single 1 into a 0.

Constraints: both height and width are in [2,20]
Movement is only EW or NS, diagonal moves not allowed
The maze is guaranteed to have a solution.

Answer: function answer() returns the length of the shortest path.
Algorithm: Represent the maze as a graph with weighted edges.
Each passable hallway is a node.  Passable locations are 
connected with edges.  If 2 areas could be connected
by removing a wall, then draw an edge, but give it a weight 
to discourage its use.